+++ internal/port/in/vote_use_case.go
@@ -4,11 +4,11 @@ import "context"
 
 type VoteUseCase interface {
 	// TODO update return type to int64
-	Vote(ctx context.Context, businessId string, messageId string, userId string) (int, error)
+	Vote(ctx context.Context, businessId string, messageId string, userId string) (int64, error)
 
-	UnVote(ctx context.Context, businessId string, messageId string, userId string) (int, error)
+	UnVote(ctx context.Context, businessId string, messageId string, userId string) (int64, error)
 
-	Count(ctx context.Context, businessId string, messageId string) (int, error)
+	Count(ctx context.Context, businessId string, messageId string) (int64, error)
 
 	IsVoted(ctx context.Context, businessId string, messageId string, userId string) (bool, error)
